% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_temps.R
\name{age_moyen_litteral}
\alias{age_moyen_litteral}
\title{Obtenir un age moyen sous forme litterale}
\usage{
age_moyen_litteral(age_moyen)
}
\arguments{
\item{format}{l'âge moyen au format numérique.}
}
\value{
l'âge moyen au format littéral.
}
\description{
Obtenir un âge moyen sous forme littérale.
}
\examples{
divr::age_moyen_literral(25.64048)

}
