% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{base_100}
\alias{base_100}
\title{Return a base 100 serie}
\usage{
base_100(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A base 100 vector
}
\description{
Return a base 100 serie
}
\examples{
x <- c(5096, 5077, 5278, 5352, 5437, 5387)

divr::base_100(x)

}
