% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charger_rdata.R
\name{charger_rdata}
\alias{charger_rdata}
\title{Charger un objet depuis un fichier RData}
\usage{
charger_rdata(rdata, nom_objet = NULL)
}
\arguments{
\item{rdata}{Le chemin vers le fichier RData.}

\item{nom_objet}{Le nom de l'objet à charger.}
}
\value{
Si le paramètre \code{nom_objet} reste à \code{NULL} alors l'ensemble des objets du fichier RData sont chargés dans une liste nommée.\cr
Sinon, seul l'élément passé en paramètre \code{nom_objet} est chargé.
}
\description{
Charger un objet depuis un fichier RData.
}
\examples{
# Chargement dans une liste de tous les objets d'un fichier RData (ici, un seul objet nommé "data_lib_num_mois"))
divr::charger_rdata(paste0(racine_packages, "divr/data/data_lib_num_mois.RData"))

# Chargement d'un objet déterminé
divr::charger_rdata(paste0(racine_packages, "divr/data/data_lib_num_mois.RData"), "data_lib_num_mois")

}

