% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{chrome_print_no_gpu}
\alias{chrome_print_no_gpu}
\title{Convert an HTML file to PDF using Google Chrome}
\usage{
chrome_print_no_gpu(html_file, pdf_file = NULL, remove_html = FALSE)
}
\arguments{
\item{html_file}{\dots}

\item{pdf_file}{\dots}

\item{remove_html}{\dots}
}
\description{
Convert an HTML file to PDF using Google Chrome
}
