% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{cl_initialise}
\alias{cl_initialise}
\title{Initialise cores for a parallel computation.}
\usage{
cl_initialise(n_cores = NULL)
}
\arguments{
\item{n_cores}{Number of cores to use (maximum by default).}
}
\value{
An objectof type cluster.
}
\description{
Initialise cores for a parallel computation.
}
