% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{cl_stop}
\alias{cl_stop}
\title{Stop a cluster after a parallel computation.}
\usage{
cl_stop(cluster)
}
\arguments{
\item{cluster}{A cluster object returnes by \code{initialise_cluster}}
}
\description{
Stop a cluster after a parallel computation.
}
