% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{code_find}
\alias{code_find}
\title{Look for code inside R or Rmd scripts.}
\usage{
code_find(code, path = getwd())
}
\arguments{
\item{code}{R syntax to search.}

\item{path}{Path where R and Rmd files are located.}
}
\value{
A data frame with three fields : file, line number and code containing researched syntax.
}
\description{
Look for code inside R or Rmd scripts.
}
