% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{code_replace}
\alias{code_replace}
\title{Look and replace code inside R or Rmd scripts.}
\usage{
code_replace(code, replacement, path)
}
\arguments{
\item{code}{R syntax to search and replace.}

\item{replacement}{A character of replacement.}

\item{path}{Path where R and Rmd files are located.}
}
\description{
Look and replace code inside R or Rmd scripts.
}
