% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{concatener_fichiers}
\alias{concatener_fichiers}
\title{Concatener des fichiers}
\usage{
concatener_fichiers(fichiers_entree, fichier_sortie)
}
\arguments{
\item{fichiers_entree}{Vecteur de fichiers à concaténer.}

\item{fichier_sortie}{Chemin du fichier en sortie.}
}
\description{
Concaténer des fichiers.
}
\examples{

}
