% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{construction_package}
\alias{construction_package}
\title{Construire un package}
\usage{
construction_package(repertoire_package)
}
\arguments{
\item{repertoire_package}{Le répertoire racine du package.}
}
\description{
Construire un package.
}
