% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_temps.R
\name{conv_lib_num_mois}
\alias{conv_lib_num_mois}
\title{Obtenir le numero de mois a partir du libelle}
\usage{
conv_lib_num_mois(lib_mois)
}
\arguments{
\item{lib_mois}{Un vecteur de libellés de mois.}
}
\value{
Un vecteur de numéros de mois.

Jeu de données source : \code{divr::data_lib_num_mois}.\cr
Il est créé à partir de la table "Libelle_num_mois" de la base Access "Tables_ref.accdb" (projet "Temps").
}
\description{
Obtenir le numéro de mois à partir du libellé.
}
\examples{
divr::conv_lib_num_mois(c("janvier", "Février", "avril"))

}

