% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_html_to_pdf}
\alias{convert_html_to_pdf}
\title{Convert an HTML file to PDF using Google Chrome}
\usage{
convert_html_to_pdf(html_file, pdf_file = NULL, remove_html = FALSE)
}
\arguments{
\item{html_file}{\dots}

\item{pdf_file}{\dots}

\item{remove_html}{\dots}
}
\description{
Convert an HTML file to PDF using Google Chrome
}
