% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_agg}
\alias{count_agg}
\title{Count aggregates (extended version of dplyr::count).}
\usage{
count_agg(data, ..., total_value = "Total")
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to group by.}

\item{total_value}{The total value label.}
}
\value{
A tbl, grouped the same way as data.
}
\description{
Count aggregates (extended version of dplyr::count).
}
\examples{
data <- dplyr::tibble(var1 = c(1, 1, 2, 2) \%>\% as.character(), var2 = c(1, 2, 1, 2) \%>\% as.character(), var3 = c(1, 2, 2, 2) \%>\% as.character())
divr::count_agg(data, var1, var2)

}
