% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_agregat}
\alias{count_agregat}
\title{count_agregat}
\usage{
count_agregat(table, champ, valeur_total)
}
\arguments{
\item{table}{\dots}

\item{champ}{\dots}

\item{valeur_total}{\dots}
}
\value{

}
\description{
count_agregat
}
\keyword{internal}
