% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_agregats}
\alias{count_agregats}
\title{count_agregats}
\usage{
count_agregats(table, ..., valeur_total = "Total")
}
\arguments{
\item{table}{\dots}

\item{...}{\dots}

\item{valeur_total}{\dots}
}
\value{

}
\description{
count_agregats
}
