% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{creer_repertoire}
\alias{creer_repertoire}
\title{Creer un repertoire}
\usage{
creer_repertoire(repertoire)
}
\arguments{
\item{repertoire}{Chemin du répertoire à créer.}
}
\description{
Créer un répertoire.
}
\examples{
# Création du repertoire "test"
divr::creer_repertoire("test")

}

