% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_temps.R
\name{date_jour}
\alias{date_jour}
\title{Obtenir la date du jour}
\usage{
date_jour(format = "datage_fichier")
}
\arguments{
\item{format}{Le format de la date à retourner (datage_fichier, par défaut | litteral).}
}
\value{
La date du jour.
}
\description{
Obtenir la date du jour.
}
\examples{
divr::date_jour_fichier()

}
