% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_temps.R
\name{date_jour_fichier}
\alias{date_jour_fichier}
\title{Obtenir la date du jour (format yyyy_mm_jj) pour prefixer un fichier}
\usage{
date_jour_fichier()
}
\value{
La date du jour au format yyyy_mm_jj
}
\description{
Obtenir la date du jour (format yyyy_mm_jj) pour préfixer un fichier.
}
\examples{
divr::date_jour_fichier()

}

