% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dependances_package}
\alias{dependances_package}
\title{Obtenir les dependances d'un packages}
\usage{
dependances_package(repertoire = ".")
}
\arguments{
\item{repertoire}{Le répertoire racine du package.}
}
\value{
Un vecteur de dépendances.
}
\description{
Obtenir les dépendances d'un packages.
}
\examples{

}
