% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divR.R
\name{divR}
\alias{divR}
\title{Distributional IV Model Function}
\usage{
divR(
  Z,
  X,
  Y,
  W = NULL,
  epsx_dim = 50,
  epsy_dim = 50,
  epsh_dim = 50,
  hidden_dim = 100,
  num_layer = 3,
  num_epochs = 1000,
  lr = 10^(-3),
  beta = 1,
  silent = FALSE,
  standardize = TRUE
)
}
\arguments{
\item{Z}{A data frame, matrix, vector, or factor variable representing the instrumental variable.}

\item{X}{A data frame, matrix, vector, or factor variable representing the predictor.}

\item{Y}{A data frame, matrix, vector, or factor variable representing the target variable.}

\item{W}{(Optional) A data frame, matrix, vector, or factor variable representing the exogenous variable(s).}

\item{epsx_dim}{The dimension of the noise corresponding to the predictor introduced in the model (default: 50).}

\item{epsy_dim}{The dimension of the noise corresponding to the outcome introduced in the model (default: 50).}

\item{epsh_dim}{The dimension of the noise corresponding to the hidden confounder introduced in the model (default: 50).}

\item{hidden_dim}{The size of the hidden layer in the model (default: 100).}

\item{num_layer}{The number of layers in the model (default: 3).}

\item{num_epochs}{The number of epochs to be used in training (default: 1000).}

\item{lr}{The learning rate to be used in training (default: 10^-3).}

\item{beta}{The beta scaling factor for energy loss, numeric value from (0,2) (default: 1).}

\item{silent}{A boolean indicating whether to suppress output during model training (default: FALSE).}

\item{standardize}{A boolean indicating whether to standardize the input data (default: TRUE).}
}
\value{
A distributional IV model object with class 'divR'.
}
\description{
This function fits a distributional IV model to the data. It allows for
the tuning of several parameters related to model complexity and model training.
Variables are per default internally standardized (predictions are on the original scale).
}
