% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublons.R
\name{doublons}
\alias{doublons}
\title{Extraire les doublons d'une table selon un ou plusieurs champs}
\usage{
doublons(table, cle)
}
\arguments{
\item{table}{Data frame dans lequel sont extraits les doublons.}

\item{cle}{Vecteur de noms de champ au croisement desquels sont extraits les doublons.}
}
\value{
Un extrait du data frame correspondant aux enregistrements doublon.
}
\description{
Extraire les doublons d'une table selon un ou plusieurs champs.
}
\examples{
data_frame <- dplyr::data_frame(cle1 = c("A", "A", "B", "B"), cle2 = c("1", "1", "2", "3"), champ = 1:4)

# Un exemple avec des doublons
divr::doublons(data_frame, cle = c("cle1", "cle2"))

# Un exemple sans doublon
divr::doublons(data_frame, cle = c("cle1", "cle2", "champ"))

}
