% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{duplicate}
\alias{duplicate}
\title{Extract duplicate rows from a data frame.}
\usage{
duplicate(data, ...)
}
\arguments{
\item{A}{data frame.}
}
\value{
A data frame containing only the duplicate rows.
}
\description{
Extract duplicate rows from a data frame.
}
\examples{
data <- dplyr::tibble(cle1 = c("A", "A", "B", "B"), cle2 = c("1", "1", "2", "3"), champ = 1:4)

# With duplicate
divr::duplicate(data, cle1, cle2)

# Without duplicate
divr::duplicate(data, cle1, cle2, champ))

}
