% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{extension_fichier}
\alias{extension_fichier}
\title{Extraire l'extension d'un fichier}
\usage{
extension_fichier(fichier)
}
\arguments{
\item{fichier}{Chemin du fichier dont l'extension est extraite.}
}
\value{
L'extension du fichier.
}
\description{
Extraire l'extension d'un fichier.
}
\examples{
divr::extension_fichier(c("fichier_excel.xlsx", "fichier_word.docx"))

}

