% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{extraire_chemin}
\alias{extraire_chemin}
\title{Extraire le chemin a partir du chemin complet}
\usage{
extraire_chemin(chemin_fichier)
}
\arguments{
\item{chemin_fichier}{Chemin complet dont le chemin est extrait.}
}
\value{
Le chemin.
}
\description{
Extraire le chemin à partir du chemin complet.
}
\examples{
divr::extraire_chemin(c("Test/Dossier/fichier_excel.xlsx", "Test2/Dossier2/fichier_word.docx"))

}
