% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{extraire_fichiers_zip}
\alias{extraire_fichiers_zip}
\title{Extraire des fichiers d'une archive zip}
\usage{
extraire_fichiers_zip(archive_zip, regex_fichier = NULL,
  repertoire_sortie = ".", supprimer_zip = FALSE)
}
\arguments{
\item{archive_zip}{Chemin du fichier zip dont le contenu est extrait.}

\item{regex_fichier}{Expression régulière pour filtrer les fichiers à extraire.}

\item{repertoire_sortie}{Chemin du répertoire vers lequel est réalisée l'extraction (par défaut, le répertoire de travail).}

\item{supprimer_zip}{\code{TRUE}: l'archive zip est supprimée; \code{FALSE}: l'archive zip n'est pas supprimée..}
}
\description{
Extraire des fichiers d'une archive zip.
}
\examples{
# Extraire les fichiers pdf d'une archive
divr::extraire_fichiers_zip("archive.zip", regex_fichier = "pdf$", repertoire_sortie = "test_zip")

}

