% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{extraire_masse_zip}
\alias{extraire_masse_zip}
\title{Extraire les fichiers des archives zip d'un repertoire (recursif)}
\usage{
extraire_masse_zip(chemin, regex_fichier, regex_zip = NULL,
  return_tibble = TRUE, paralleliser = FALSE)
}
\arguments{
\item{chemin}{Répertoire dans lequel sont cherchés (récursif) les archives zip dont le contenu est extrait.}

\item{regex_fichier}{Expression régulière pour filtrer les fichiers csv à extraire.}

\item{regex_zip}{Expression régulière pour filtrer les archives zip à traiter.}

\item{return_tibble}{Retourne ou non un data frame fournissant la correspondance entre les archives zip initiales et l'emplacement des fichiers extraits.}

\item{paralleliser}{\code{TRUE}, extraction parallelisée des archives zip.}
}
\description{
Extraire des fichiers des archives zip d'un répertoire (récursif).
}
\examples{
# Extraire les fichiers pdf stockés dans les archives zip du répertoire "Chemin/vers/un/répartoire"
divr::extraire_masse_zip("Chemin/vers/un/répartoire", regex_fichier = "pdf$", repertoire_sortie = "test_zip")

}
