% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{extraire_nom_fichier}
\alias{extraire_nom_fichier}
\title{Extraire le fichier a partir du chemin complet}
\usage{
extraire_nom_fichier(chemin_fichier)
}
\arguments{
\item{chemin_fichier}{Chemin complet dont le fichier est extrait.}
}
\value{
Le nom du fichier.
}
\description{
Extraire le fichier à partir du chemin complet.
}
\examples{
divr::extraire_nom_fichier(c("Dossier/fichier_excel.xlsx", "Dossier2/fichier_word.docx"))

}
