% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{files_append}
\alias{files_append}
\title{Append multiples files into one.}
\usage{
files_append(input_files, output_file)
}
\arguments{
\item{input_files}{A vector of files to append.}

\item{output_file}{An output file.}
}
\description{
Append multiples files into one.
}
