% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw.R
\name{\%btw\%}
\alias{\%btw\%}
\alias{\%sbtw\%}
\title{Operator to see if value between values of vector}
\usage{
x \%btw\% vec

x \%sbtw\% vec
}
\arguments{
\item{x}{a numeric value or vector}

\item{vec}{a numeric vector}
}
\description{
Custom operator that takes a single value or a numeric vector and returns \code{TRUE}
if values between minimum and maximum of a given vector.
}
\examples{

my_vec <- c(1, 4, 8)

1 \%btw\% my_vec  # returns TRUE
1.3 \%btw\% my_vec  # returns TRUE
0 \%btw\% my_vec  # returns FALSE
c(0,1.3,8, 12) \%sbtw\% my_vec # returns FALSE TRUE FALSE FALSE

}
