% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.R
\name{\%within\%}
\alias{\%within\%}
\title{Operator to see if value between values of vector}
\usage{
x \%within\% vec
}
\arguments{
\item{x}{a number}

\item{vec}{a numeric vector}
}
\description{
Custom operator that takes a value and a numeric vector, returns \code{TRUE}
if values between minimum and maximum of vector.
}
\examples{

my_vec <- c(1, 4, 8)

1 \%within\% my_vec  # should return TRUE
1.3 \%within\% my_vec  # should return TRUE
0 \%within\% my_vec  # should return FALSE

}

