% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelr.R
\name{initialiser_cluster}
\alias{initialiser_cluster}
\title{Initialiser les coeurs pour un traitement parallelise}
\usage{
initialiser_cluster(n_coeurs = NULL)
}
\arguments{
\item{n_coeurs}{Nombre de coeurs à utiliser (le maximum par défaut).}
}
\value{
Un objet de type cluster.\cr
Après les calculs parallèles, il convient de libérer le cluster avec la commande \code{parallel::stopCluster()} en le passant en paramètre.
}
\description{
Initialiser les coeurs pour un traitement parallélisé.
}
\examples{
# Créer le cluster:
cluster <- divr::initialiser_cluster()

# Un traitement parallélisé

# A l'issue du traitement parallélisé, libérer les coeurs:
divr::stopCluster(cluster)

}

