% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LES}
\alias{LES}
\title{LES dataset}
\format{A data frame with 2548 rows and 21 variables:
\describe{
  \item{Code}{description}
  \item{Dataset}{description}
  \item{BIOME}{description}
  \item{Species}{description}
  \item{GF}{description}
  \item{Decid_Egreen}{description}
  \item{Needle_Broadlf}{description}
  \item{C3C4}{description}
  \item{N2_fixer}{description}
  \item{log_LL}{description}
  \item{log_LMA}{description}
  \item{log_Nmass}{description}
  \item{log_Narea}{description}
  \item{log_Pmass}{description}
  \item{log_Parea}{description}
  \item{log_Amass}{description}
  \item{log_Aarea}{description}
  \item{log_Gs}{description}
  \item{log_Rdmass}{description}
  \item{log_Rdarea}{description}
  \item{Ca_Ci}{description}

}}
\usage{
LES
}
\description{
A dataset containing measures of LES.
}
\keyword{datasets}
