% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot3DsimiliWright2004.R
\name{LESPlot3D}
\alias{LESPlot3D}
\title{LESPlot3D}
\usage{
LESPlot3D(yourdata, Xaxis, Yaxis, Zaxis, xlab = "X", ylab = "Y",
  zlab = "Z", Xticks = NULL, Yticks = NULL, Zticks = NULL,
  logxyz = "", Rotate = 110, Flip = 25)
}
\arguments{
\item{yourdata}{name of your datasheet.}

\item{Xaxis, Yaxis, Zaxis}{Variable names of your datasheet in quotes.}

\item{xlab, ylab, zlab}{labels for axes.}

\item{Xticks, Yticks, Zticks}{a vector giving the ticks to use on the axes.}

\item{logxyz}{a character string which contains "x" if the x axis is to be logarithmic, "y" if the y axis is to be logarithmic and "z" if the axis is to be logarithmic.}

\item{Rotate, Flip}{the angles defining the viewing direction. Rotate gives the azimuthal direction and Flip the colatitude. Respectively, theta and Phi in persp.}
}
\description{
Produce 3D Plot designed as Wrigth 2004 paper
}
\examples{
data("LES")
LESPlot3D(LES,"log_LL","log_LMA","log_Nmass","LL","LMA","Nmass")
}
