% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdata.R
\name{load_rdata}
\alias{load_rdata}
\title{Load an objet from a RData file.}
\usage{
load_rdata(rdata, object = NULL)
}
\arguments{
\item{rdata}{RData file path.}

\item{object}{Object name to load.}
}
\value{
If parameter \code{object} stays \code{NULL} then all objects of the RData file loaded in a named list.\cr
Otherwise, only the object specified is loaded.
}
\description{
Load an objet from a RData file.
}
\examples{
divr::load_rdata(paste0(racine_packages, "divr/data/load_rdata.RData"))
divr::load_rdata(paste0(racine_packages, "divr/data/load_rdata.RData"), "lib_num_mois")

}
