% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{maj_champ}
\alias{maj_champ}
\title{maj_champ}
\usage{
maj_champ(table, table_pivot, champ_maj, ..., doublons = TRUE)
}
\arguments{
\item{table}{\dots}

\item{table_pivot}{\dots}

\item{champ_maj}{\dots}

\item{doublons}{\dots}
}
\description{
maj_champ
}
