% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_quantile_density.R
\name{make_quantile_density}
\alias{make_quantile_density}
\title{Make density plot with quantiles}
\usage{
make_quantile_density(ex_vector, probs = NULL)
}
\arguments{
\item{ex_vector}{a numeric vector from which the density should be computed}

\item{probs}{a numeric vector giving the quantile (default: \code{NULL} but
from 0 to 1 by step of 0.1)}
}
\description{
Plot the density of a numeric vector and show the different quantiles colored
in areas
}
\examples{

vec = rnorm(1000)

gg = make_quantile_density(vec)

gg

}
