% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhead.R
\name{mhead}
\alias{mhead}
\title{Head of a matrix}
\usage{
mhead(given_object, n = 5)
}
\arguments{
\item{given_object}{a \code{matrix} object}

\item{n}{an integer number of the number of rows and columns to show}
}
\description{
This function makes it easier to show the head of a matrix object without
all the columns. It takes a matrix object and a number and show the
given number of rows and columns
}
\examples{
m <- matrix(rnorm(1000), ncol = 1000)
mhead(m)

}
