% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{mongo_daemon}
\alias{mongo_daemon}
\title{Lancer le daemon Mongo a partir de la base}
\usage{
mongo_daemon(chemin_base_mongo)
}
\arguments{
\item{chemin_base_mongo}{Le chemin vers la base de données Mongo.}
}
\description{
Lancer le daemon Mongo à partir de la base
}
\examples{
# Le dernier dossier du répertoire "_traitement/unzip" (Il s'agit du répertoire où sont dézippés les backups Mongo)
paste0(disque, "_traitement/unzip") \%>\%
  list.dirs(recursive = FALSE) \%>\%
  tail(1) \%>\%
  divr::mongo_daemon()

}

