% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{package_build}
\alias{package_build}
\title{Build a package.}
\usage{
package_build(package_path, documentation = TRUE)
}
\arguments{
\item{package_path}{Package path.}

\item{documentation}{If \code{TRUE}, documentation is also generated.}
}
\description{
Build a package.
}
