% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{package_deps}
\alias{package_deps}
\title{Get package dependancies.}
\usage{
package_deps(package_path = getwd())
}
\arguments{
\item{package_path}{Package path.}
}
\value{
A dependencies vector.
}
\description{
Get package dependancies.
}
