% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{patch_vector}
\alias{patch_vector}
\title{Patch a current vector from a target vector.}
\usage{
patch_vector(current, target, only_na = FALSE)
}
\arguments{
\item{current}{A vector to be patched.}

\item{target}{A patch vector. If length 1 then recycled to the length of current and only NA are patched.}

\item{only_na}{If \code{TRUE} then only missing values in current are patched.}
}
\value{
A patched vector.
}
\description{
Patch a current vector from a target vector.
}
\examples{
divr::patch_vector(c(1, NA_real_, 3), c(4, 5, 6))
divr::patch_vector(c(1, NA_real_, 3), c(4, 5, 6), only_na = TRUE)
divr::patch_vector(c(1, NA_real_, 3), 4)

}
