% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plot.pca}
\alias{plot.pca}
\title{Plot method for \code{pca} objects}
\usage{
\method{plot}{pca}(pca, axes = c(1, 2), xlim = c(-1.2, 1.2),
  ylim = c(-1.2, 1.2), ind = FALSE, pch = 21, bg = "white",
  cex.p = 1, las = 1, var.labels = NULL, ...)
}
\arguments{
\item{pca}{a \code{pca} object from the \code{dudi.pca} function in \code{ade4}}

\item{axes}{a length 2 vector specifying the components to plot}

\item{xlim, ylim}{numeric vectors of length 2, giving the x and y coordinates ranges}

\item{ind}{logical value to project individuals in the correlation circle using standardized axes}

\item{pch}{symbole code for individuals points if \code{ind} is TRUE}

\item{bg}{color used for individuals points if \code{ind} is TRUE}

\item{cex.p}{individuals points size if \code{ind} is TRUE}

\item{las}{text orientation on axes}

\item{var.labels}{character string or expression used for variable names}
}
\description{
Custom function to plot \code{pca} objects from \code{ade4} package
}
