% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOmbro.R
\name{plotOmbro}
\alias{plotOmbro}
\title{climate plots}
\usage{
plotOmbro(temp, prec, min = NA, max = NA, site = NA, alt = NA,
  Tmin = NA, H = NA, cex = 1)
}
\arguments{
\item{temp}{a numeric vector monthly mean temperatures (12 values)}

\item{prec}{a numeric vector monthly total precipitations (12 values)}

\item{min}{an numeric vector montly minimal temperatures (12 values); optional, default value = NA}

\item{max}{a numeric vector montly maximal temperatures (12 values); optional, default value = NA}

\item{site}{the site name character}

\item{alt}{the site altitude; optional, default value = NA}

\item{Tmin}{minimal value for the temperature y-axis; default value = NA}

\item{H}{a two colomn matrix describing the low water availability periods (start and end between 0 and 12); optional, default value = NA}

\item{cex}{text size; default value = 1}
}
\description{
Produce a climate plot from monthly mean temperature and total precipitations
}
\examples{

temp<-c(6.5,7.2,9.9,12.5,16.5,20.6,23.8,23.4,19.4,15.4,10.3,7.3)
min<-c(1.7,2,4.5,7.3,11.4,14.9,17.9,17.5,13.8,10.7,5.4,2.5)
max<-c(11.4,12.2,15.3,17.8,21.7,26.2,29.6,29.3,25.1,20.1,14.9,11.9)
prec<-c(67.9,64.1,40.6,66.3,54.1,35.4,20.4,40.3,105.1,113.9,85.7,87.7)
par(mar=c(3,3,1,3))
plotOmbro(temp,prec,min,max,site="Montpellier",alt=57)


}

