% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.divR.R
\name{predict.divR}
\alias{predict.divR}
\title{Prediction Function for a divR Model Object}
\usage{
\method{predict}{divR}(
  object,
  Xtest,
  Wtest = NULL,
  type = c("mean", "sample", "quantile")[1],
  trim = 0.05,
  quantiles = 0.1 * (1:9),
  nsample = 200,
  drop = TRUE,
  ...
)
}
\arguments{
\item{object}{A trained divR model returned from \code{\link{divR}}.}

\item{Xtest}{A matrix or data frame representing predictors in the test set.}

\item{Wtest}{A matrix or data frame representing exogenous predictors in the test set.}

\item{type}{The type of prediction to make: "mean", "sample", or "quantile".}

\item{trim}{The proportion of extreme values to trim when calculating the mean (default: 0.05).}

\item{quantiles}{The quantiles to estimate if type is "quantile" (default: 0.1*(1:9)).}

\item{nsample}{The number of samples to draw (default: 200).}

\item{drop}{A boolean indicating whether to drop dimensions of length 1 from the output (default: TRUE).}

\item{...}{additional arguments (currently ignored).}
}
\value{
A vector or matrix/array of predictions.
}
\description{
This function computes predictions from a trained divR model. It allows for estimation
of the interventional mean and quantiles, as well as sampling from the fitted interventional distribution.
}
