% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.divR.R
\name{print.divR}
\alias{print.divR}
\title{Print Function for a divR Model Object}
\usage{
\method{print}{divR}(x, ...)
}
\arguments{
\item{x}{A trained divR model returned from \code{\link{divR}}.}

\item{...}{additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object. Prints a summary of the model
architecture, training process, and loss values.
}
\description{
This function displays a summary of a fitted divR model object.
}
