% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{rechercher_code}
\alias{rechercher_code}
\title{Rechercher une syntaxe de code dans les fichiers R}
\usage{
rechercher_code(code, chemin)
}
\arguments{
\item{code}{Le code R à rechercher.}

\item{chemin}{Le chemin contenant les programmes R.}
}
\value{
Un tibble à trois champs : le fichier, le numéro de ligne et le code contenant la synatxe recherchée.
}
\description{
Rechercher une syntaxe de code dans les fichiers R.
}
