% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{remplacer_code}
\alias{remplacer_code}
\title{Remplacer une syntaxe de code dans les fichiers R}
\usage{
remplacer_code(code, remplacement, chemin)
}
\arguments{
\item{code}{Le code R à rechercher.}

\item{remplacement}{Le code R de remplacement.}

\item{chemin}{Le chemin contenant les programmes R.}
}
\description{
Remplacer une syntaxe de code dans les fichiers R.
}
