% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{remplacer_si}
\alias{remplacer_si}
\title{remplacer_si}
\usage{
remplacer_si(courant, cible, condition = TRUE, na_cible = FALSE)
}
\arguments{
\item{courant}{\dots}

\item{cible}{\dots}

\item{condition}{\dots}

\item{na_cible}{\dots}
}
\description{
remplacer_si
}
