% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{round_100}
\alias{round_100}
\title{Round with a sum 100 (avoid round errors)}
\usage{
round_100(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A round numeric vector
}
\description{
Round with a sum 100 (avoid round errors)
}
\examples{
x <- c(81.808219, 6.575342, 11.616438)

sum(x)
sum(round(x))

divr::round_100(x)
sum(divr::round_100(x))

}
