% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Stop a cluster after a parallel computation.}
\usage{
stop_cluster(cluster)
}
\arguments{
\item{cluster}{A cluster object returnes by \code{initialise_cluster}}
}
\description{
Stop a cluster after a parallel computation.
}
