% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelr.R
\name{stopper_cluster}
\alias{stopper_cluster}
\title{Stopper un cluster (traitement parallelise)}
\usage{
stopper_cluster(cluster)
}
\arguments{
\item{cluster}{}
}
\description{
Stopper un cluster (traitement parallélisé)
}
\examples{
#' # Créer le cluster:
cluster <- divr::initialiser_cluster()

# Un traitement parallélisé

# A l'issue du traitement parallélisé, libérer les coeurs:
divr::stopper_cluster(cluster)

}

