% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{supprimer_repertoire}
\alias{supprimer_repertoire}
\title{Supprimer un repertoire}
\usage{
supprimer_repertoire(repertoire)
}
\arguments{
\item{repertoire}{Chemin du répertoire à supprimer.}
}
\description{
Supprimer un repertoire.
}
\examples{
# Création du répertoire "test"
divr::creer_repertoire("test")

# Suppression du répertoire
divr::supprimer_repertoire("test")

}

