% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{vider_repertoire}
\alias{vider_repertoire}
\title{Vider un repertoire}
\usage{
vider_repertoire(repertoire)
}
\arguments{
\item{repertoire}{Chemin du répertoire à vider.}
}
\description{
Vider un repertoire.
}
\examples{
# Création du repertoire "test"
divr::creer_repertoire("test")

# Ecriture d'un fichier à l'intérieur
write.csv(1, "test/test.csv")

# Vidage du répertoire
divr::vider_repertoire("test")

}
