% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplot.R
\name{vioplot}
\alias{vioplot}
\title{Violon plots}
\usage{
vioplot(x, weights, ...)
}
\arguments{
\item{x}{a numeric vector of data}

\item{weights}{a numeric vector with the same length as x}

\item{wex}{the maximum value for the density curve}

\item{h}{the kernel density smoothing parameter}

\item{add}{if TRUE adds the plot to the current plot window}

\item{nblim}{minimum length of x for computing the kernel density}

\item{boxplot}{logical value for adding a boxplot or not}

\item{side}{a character value describind if the distribution should be drawn in the "above", "below", "right", "left" or "both" side(s)}

\item{fill}{the color of the distribution area}

\item{pts}{logical value for adding x data on the curve distrivution}

\item{boxplot.fill}{the color of the boxplot}

\item{horizontal}{logical value}
}
\description{
Produce a violon plot taking into acount weightings
}
\examples{

data("ChickWeight")
vioplot(x = chickwts$weight, horizontal = FALSE, side="both", fill="grey70", boxplot.fill="grey30", label = "Weights", xlab = "none")

}
\seealso{
For group comparison see vioplotGp
}

