% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplotGP.R
\name{vioplotGP}
\alias{vioplotGP}
\title{Violon plots for Groups}
\usage{
vioplot(x, gp, weights, ...)
}
\arguments{
\item{x}{a numeric vector of data}

\item{gp}{factors with the same length as x decribing groups}

\item{weights}{a numeric vector with the same length as x}

\item{wex}{the maximum value for the density curve}

\item{h}{the kernel density smoothing parameter}

\item{nblim}{minimum length of x for computing the kernel density}

\item{Test}{a character value for group comparison tests "tukey" or "wilcox"}

\item{add}{if TRUE adds the plot to the current plot window}

\item{boxplot}{logical value for adding a boxplot or not}

\item{side}{a character value describind if the distribution should be drawn in the "above", "below", "right", "left" or "both" side(s)}

\item{fill}{the color of the distribution area}

\item{pts}{logical value for adding x data on the curve distrivution}

\item{boxplot.fill}{the color of the boxplot}

\item{horizontal}{logical value}
}
\description{
Produce a violon plot acording to groups taking into acount weightings and compute comparison tests
}
\examples{

data("ChickWeight")
vioplotGP(x = chickwts$weight, gp= chickwts$feed, las=2, horizontal = F, side="both",
wex=.4, fill="grey70", boxplot.fill="grey30", xlab="Chick weights (mg)", Test="tukey",
pos.letters = 2, col.letters = "white")

}
\seealso{
For group comparison see vioplotGp
}

