% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{zip_extract_file}
\alias{zip_extract_file}
\title{Extract files from a zip.}
\usage{
zip_extract_file(zip_file, pattern = NULL, exdir = NULL,
  remove_zip = FALSE)
}
\arguments{
\item{zip_file}{Path to the zip file.}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be extracted.}

\item{exdir}{The directory to extract files to. It will be created if necessary.}

\item{remove_zip}{If \code{TRUE}, the zip file is removed.}
}
\description{
Files within the zip file can be selected with the pattern argument.
}
\examples{
divr::zip_extract_file("archive.zip", pattern = "pdf$", exdir = "test_zip")

}
