% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_fichiers.R
\name{zip_extract_path}
\alias{zip_extract_path}
\title{Extract files from zip located in a path.}
\usage{
zip_extract_path(path, pattern, pattern_zip = "\\\\.zip$",
  n_files = Inf, parallel = FALSE)
}
\arguments{
\item{path}{Path where the zip files are located (recursive).}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be extracted.}

\item{pattern_zip}{an optional regular expression. Only zip file names which match the regular expression will be unziped.}

\item{n_files}{Number of files to extract. A negative value will starts from the bottom of the files list.}

\item{parallel}{If \code{TRUE}, a parallelised extraction is performed.}
}
\description{
Extract files from zip located in a path.
}
\examples{
divr::zip_extract_path("Chemin/vers/un/répartoire", pattern = "pdf$")

}
