% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{construction_package}
\alias{construction_package}
\title{Construire un package}
\usage{
construction_package(package, documentation = TRUE)
}
\arguments{
\item{package}{Le nom du package.}

\item{documentation}{Génération de la documentation, \code{TRUE} par défaut.}
}
\description{
Construire un package.
}
