% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_nwk.R
\name{cor_nwk}
\alias{cor_nwk}
\title{Plot a correlation diagram}
\usage{
cor_nwk(dat, labels = NA, method = "pearson", order = TRUE, col = c(1,
  "grey80"))
}
\arguments{
\item{dat}{a matrix or data.frame with variables in column}

\item{labels}{custom label vector in the same order than column names}

\item{method}{method used for the correlation test}

\item{order}{logical to order according to a clustering method}

\item{col}{2 length vector of color names for positive and negative correlation}
}
\description{
Summary diagram of the correlations between variables
}
\examples{

data("mtcars")
cor_nwk(mtcars)

}
\author{
Jules Segrestin
}
