% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_nwk.R
\name{cor_nwk}
\alias{cor_nwk}
\title{Plot a correlation diagram}
\usage{
cor_nwk(dat, labels = NA, phylo = NULL, method = "pearson",
  p.critic = 0.05, pval.adjust = NULL, order = TRUE, col = c(1,
  "grey50"), lty = c(1, 1))
}
\arguments{
\item{dat}{a matrix or data.frame with variables in column.}

\item{labels}{custom label vector in the same order than column names.}

\item{phylo}{NULL or an object of class \code{phylo} representing the phylogeny (with branch lengths) to consider}

\item{method}{method used for the correlation test. One of "pearson" (default), "kendall", or "spearman".}

\item{p.critic}{Critical p.value under which a correlation line should be drawn}

\item{pval.adjust}{method to adjust p-values in "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr".}

\item{order}{logical to order variables according to a clustering method.}

\item{col}{2 length vector of colour names for positive and negative correlation.}

\item{lty}{2 length vector of line type for positive and negative correlation (see \code{par} for more details).}
}
\description{
Summary diagram of the correlations between variables
}
\examples{
data("mtcars")
cor_nwk(mtcars)

}
\author{
Jules Segrestin
}
