% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_agregats}
\alias{count_agregats}
\title{count_agregats}
\usage{
count_agregats(table, ..., valeur_total = "Total")
}
\arguments{
\item{table}{\dots}

\item{...}{\dots}

\item{valeur_total}{\dots}
}
\value{

}
\description{
count_agregats
}
\examples{
dplyr::tibble(var1 = c(1, 1, 2, 2) \%>\% as.character(), var2 = c(1, 2, 1, 2) \%>\% as.character(), var3 = c(1, 2, 2, 2) \%>\% as.character())

}
