% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot3DsimiliWright2004.R
\name{LESPlot3D}
\alias{LESPlot3D}
\title{LES Plot3D}
\usage{
LESPlot3D(yourdata, Xaxis, Yaxis, Zaxis, xlab = "X", ylab = "Y",
  zlab = "Z", logxyz = "", Rotate = 110, Flip = 25)
}
\arguments{
\item{yourdata}{a datasheet containing the variables to plot.}

\item{Xaxis}{name of the variable to plot in X}

\item{xlab}{label for x axis}

\item{logxyz}{character containing the axes to log (log10)}

\item{Rotate}{parameter to moove the plot around a vertical axe}

\item{Flip}{parameter to moove the plot around a horizontal axe}
}
\description{
Produce the same 3D Plot as found in Wrigth 2004 paper.
}
\examples{

LESPlot3D(iris, "Sepal.Length", "Sepal.Width", "Petal.Length",
          "Sepal.Length", "Sepal.Width", "Petal.Length")
}
