% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_temps.R
\name{mois_ecoules}
\alias{mois_ecoules}
\title{Obtenir le nombre de mois ecoules entre deux dates}
\usage{
mois_ecoules(date_debut, date_fin)
}
\arguments{
\item{date_debut}{Date de début de la période}

\item{date_fin}{Date de fin de la période}
}
\value{
Le nombre de mois écoulés entre les deux dates.\cr
Un seul jour dans un mois suffit à le comptabiliser.
}
\description{
Obtenir le nombre de mois écoulés entre deux dates.
}
\examples{
# Une année écoulée au jour près
divr::mois_ecoules("2000-01-01", "2001-01-01")

# Une année écoulée moins un jour
divr::mois_ecoules("2000-01-01", "2000-12-31")

}

