% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplotGP.R
\name{vioplotGP}
\alias{vioplotGP}
\title{Violin plots for Groups}
\usage{
vioplotGP(x, gp, weights = NULL, labels = NA, xlab = NA, xlim = NA,
  ylim = NA, at = NA, wex = 0.75, h = NA, nblim = 10, Test = NA,
  adj.Test = 0.01, line.Test = NA, side.Test = 3, Fname = NA,
  type.letters = "latin", pos.letters = 1, col.letters = 1, add = F,
  col = 1, pch = 19, lwd = 1, lty = 1, cex = 1, cex.points = 1,
  col.points = 1, bg = 1, las = 0, boxplot = TRUE, side = "above",
  fill = NA, pts = TRUE, boxplot.fill = "white", horizontal = TRUE)
}
\arguments{
\item{x}{a numeric vector of data}

\item{gp}{factors with the same length as x describing groups}

\item{weights}{a numeric vector with the same length as x}

\item{labels}{group names in the same order than \code{gp} levels}

\item{xlab}{label for the x-axis}

\item{xlim, }{ylim range of values used for x and y-axes}

\item{at}{position of each group in the y-axis}

\item{wex}{the maximum value for the density curve}

\item{h}{the kernel density smoothing parameter}

\item{nblim}{minimum length of x for computing the kernel density}

\item{Test}{a character value for group comparison tests "tukey" or "wilcox"}

\item{adj.Test, side.Test, line.Test}{position paramters for the test value and P.value}

\item{Fname}{?}

\item{type.letters}{if Test is done, "latin" or "greek" letters are used to show significant differences between groups}

\item{pos.letters, col.letters}{position (1, 2, 3 or 4) and colors for letters}

\item{add}{if TRUE adds the plot to the current plot window}

\item{col}{?}

\item{pch}{?}

\item{lwd}{?}

\item{lty}{?}

\item{cex}{?}

\item{cex.points}{?}

\item{col.points}{?}

\item{bg}{?}

\item{las}{?}

\item{boxplot}{logical value for adding a boxplot or not}

\item{side}{a character value describing if the distribution should be drawn
in the "above", "below", "right", "left" or "both" side(s)}

\item{fill}{the color of the distribution area}

\item{pts}{logical value for adding x data on the distribution curve}

\item{boxplot.fill}{the color of the boxplot}

\item{horizontal}{logical value}
}
\description{
Produce a Violin plot according to groups taking into account weightings and compute comparison tests
}
\examples{

data("ChickWeight")
vioplotGP(x = chickwts$weight, gp = chickwts$feed, las = 2,
          horizontal = FALSE, side = "both", wex = .4, fill = "grey70",
          boxplot.fill = "grey30", xlab = "Chick weights (mg)",
          Test = "tukey", pos.letters = 3, col.letters = "white",
          col.points = "grey50")

}
\seealso{
For group comparison see vioplotGp
}
